Move agent-browser execution into orchestration phases
