+++ test/agent-browser.extension-input-modes.test.ts
@@ -477,4 +477,3 @@ process.stdin.on("end", () => {
 		await rm(tempDir, { force: true, recursive: true });
 	}
 });
-
