+++ test/agent-browser.extension-source-lookup.test.ts
@@ -383,4 +383,3 @@ process.stdin.on("end", () => {
 		await rm(tempDir, { force: true, recursive: true });
 	}
 });
-
